desktop_server/requirements.txt
# Project Vulcan - Desktop Control Server Dependencies
# Install: pip install -r requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Desktop Automation
pyautogui==0.9.54
pywin32==306

# Image Processing
pillow==10.2.0
mss==9.0.1

# OCR (requires Tesseract installed separately)
pytesseract==0.3.10

# Configuration
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7

# Vector Memory & RAG
chromadb==0.4.22
sentence-transformers==2.2.2

# MCP
mcp==1.0.0

# Visual Replay
opencv-python-headless==4.9.0.80

# LLM
anthropic==0.18.1

# Monitoring
psutil==5.9.8
